cheaper 8-bit integer approximation

+++ fw/src/mode_wall.h
@@ -2,9 +2,9 @@
 
 #include <stdint.h>
 
-extern uint16_t distance_left;
-extern uint16_t distance_center;
-extern uint16_t distance_right;
+extern uint8_t distance_left;
+extern uint8_t distance_center;
+extern uint8_t distance_right;
 
 void mode_wall_enter();
 void mode_wall_tick();
